Line 87 should be changed.
Hello, My name is Young from South Korea. 
My friend and I tested this code in order to deploy Azure Arc enabled App Services. There is an issue that the Powershell command doesn't work to implement namespace properly, so we spend a lot of times and found that a parameter in line 87 is incorrect.
Please Check if line 87 is correct
Thank you.
@ParkGaonNuri